+++ packages/core/expressions/src/types.ts
@@ -1,0 +1,4 @@
+import type * as monaco from 'monaco-editor'
+
+export type ProvideCompletionItems = monaco.languages.CompletionItemProvider['provideCompletionItems']
+export type ProvideRhsValueCompletion = (lhsValue: string, rhsValueValue: string, lhsRange: monaco.Range, rhsValueRange: monaco.Range) => Promise<monaco.languages.CompletionList | undefined>