+++ packages/core/forms/src/utils/isValidUuid.ts
@@ -5,8 +5,10 @@
  */
 export const uuidRegEx = '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
 
-export function isValidUuid(str: string) {
-  if (!str) return false
+export const isValidUuid = (str: string) => {
+  if (!str) {
+    return false
+  }
 
   return str.length === 36 && new RegExp(`^${uuidRegEx}$`).test(str)
 }